Refactor sightMarks structure for better component organization.
Moved sightMarks components, hooks, and styles to a consolidated folder under `components/sightMarks`. Updated imports across the app to reflect the new paths, improving maintainability and clarity of the codebase.
