--- app/(tabs)/sightMarks/components/index.ts
@@ -1,3 +1,0 @@
-export { default as ConfirmRemoveMarks } from './confirmRemoveMarks/ConfirmRemoveMarks';
-export { default as MarksForm } from './marksForm/MarksForm';
-export { default as MarksTable } from './marksTable/MarksTable';
