+++ components/sightMarks/SightMarksStyles.ts
