+++ components/sightMarks/calculateMarksModal/CalculateMarksModalStyles.ts
