+++ components/sightMarks/calculatedMarksTable/CalculatedMarksTable.tsx
