+++ components/sightMarks/calculatedMarksTable/CalculatedMarksTableStyles.ts
