+++ components/sightMarks/confirmRemoveMarks/ConfirmRemoveMarks.tsx
