+++ components/sightMarks/confirmRemoveMarks/ConfirmRemoveMarksStyles.ts
