+++ components/sightMarks/marksTable/MarksTableStyles.ts
