+++ components/sightMarks/screens/CalculateScreenStyles.ts
