+++ components/sightMarks/screens/MarksScreenStyles.ts
